Created DB file with functions that can be called when needed
